// src/app/legislator/interfaces/legislatorDetail.ts
export interface LegislatorDetail {
  data: {
    bioguideID: string;
    chamber: string;
    committees: string[];
    coordinates: number[][];
    district: string;
    facebookAccount: string;
    firstName: string;
    following: boolean;
    middleName: string;
    nextElection: string;
    lastName: string;
    location: string;
    office: string;
    party: string;
    phone: string;
    position: string;
    state: string;
    terms: string[];
    twitterAccount: string;
    url: string;
    youtubeAccount: string;
  };
}